+++ src/main/java/com/zincoid/nullbot/core/mapper/DailyMapper.java
@@ -3,11 +3,11 @@ package com.zincoid.nullbot.core.mapper;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
-import com.zincoid.nullbot.core.model.data.po.StatisticDatePO;
+import com.zincoid.nullbot.core.model.data.po.DailyPO;
 
 @Mapper
-public interface StatisticDateMapper extends BaseMapper<StatisticDatePO> {
+public interface DailyMapper extends BaseMapper<DailyPO> {
 
-    @Select("SELECT COALESCE(SUM(visits), 0) FROM statistic_date")
+    @Select("SELECT COALESCE(SUM(visits), 0) FROM daily")
     Long selectTotalVisits();
 }