+++ src/main/java/com/zincoid/nullbot/core/model/data/po/BottlePO.java
@@ -7,7 +7,7 @@ import com.mikuac.shiro.common.utils.MsgUtils;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import com.zincoid.nullbot.bot.command.game.single.DriftBottleCommand;
+import com.zincoid.nullbot.bot.command.game.single.BottleCommand;
 import com.zincoid.nullbot.core.util.Base64Util;
 
 import java.time.LocalDateTime;
@@ -16,8 +16,8 @@ import java.time.format.DateTimeFormatter;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@TableName("`drift_bottle`")
-public class DriftBottlePO {
+@TableName("`bottle`")
+public class BottlePO {
 
     @TableId(value = "id", type = IdType.AUTO)
     private Integer id;
@@ -44,7 +44,7 @@ public class DriftBottlePO {
             
             """
                 .formatted(
-                        DriftBottleCommand.getKeepTimeoutSeconds(),
+                        BottleCommand.getKeepTimeoutSeconds(),
                         id,
                         rethrowTimes == 0 ? "首次被捡到" : "已被投回 " + rethrowTimes + " 次",
                         formattedTime