+++ src/main/java/com/zincoid/nullbot/core/service/system/StatsService.java
@@ -1,0 +1,14 @@
+package com.zincoid.nullbot.core.service.system;
+
+import com.zincoid.nullbot.core.model.data.vo.StatsVO;
+
+public interface StatsService {
+
+    void increaseDaily();
+
+    void increase(Long groupId, Long userId, String command);
+
+    StatsVO getStatsVO();
+
+    Long getUsage(Long userId);
+}