+++ src/main/resources/mysql/stats.sql
@@ -1,16 +1,16 @@
 # CREATE DATABASE nullbot;
 # USE nullbot;
 
-DROP TABLE IF EXISTS `statistic_date`;
-DROP TABLE IF EXISTS `statistic`;
+DROP TABLE IF EXISTS `daily`;
+DROP TABLE IF EXISTS `stats`;
 
-CREATE TABLE `statistic_date` (
+CREATE TABLE `daily` (
     id INT PRIMARY KEY AUTO_INCREMENT,
     date DATE,
     visits BIGINT DEFAULT 0
 );
 
-CREATE TABLE `statistic` (
+CREATE TABLE `stats` (
     id INT PRIMARY KEY AUTO_INCREMENT,
     group_id LONG NOT NULL,
     user_id LONG NOT NULL,